+++ src/Request/DataTablesClearTableRequest.php
@@ -1,0 +1,67 @@
+<?php
+declare(strict_types=1);
+
+namespace Citilink\ExpertSenderApi\Request;
+
+use Citilink\ExpertSenderApi\Enum\HttpMethod;
+use Citilink\ExpertSenderApi\RequestInterface;
+
+/**
+ * Clear table request
+ *
+ * @author Nikita Sapogov <sapogov.n@citilink.ru>
+ */
+class DataTablesClearTableRequest implements RequestInterface
+{
+    /**
+     * @var string Table name
+     */
+    private $tableName;
+
+    /**
+     * Constructor.
+     *
+     * @param string $tableName Table name
+     */
+    public function __construct(string $tableName)
+    {
+        $this->tableName = $tableName;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function toXml(): string
+    {
+        $xmlWriter = new \XMLWriter();
+        $xmlWriter->openMemory();
+
+        $xmlWriter->writeElement('TableName', $this->tableName);
+
+        return $xmlWriter->flush();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getQueryParams(): array
+    {
+        return [];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getMethod(): HttpMethod
+    {
+        return HttpMethod::POST();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getUri(): string
+    {
+        return '/Api/DataTablesClearTable';
+    }
+}